requirements.txt
pandas
numpy
matplotlib
psutil
py-cpuinfo